NEEDSCHECK: Use os.path functions more.
Helps with OS compatibility.

Also, fix an apparent directory resolution bug in Dar_File.extractFile.
I'm not sure what the intention was.

Also, afs's and gmp's extractFile have a different way of inserting the
offsert into the filename, which work differently if the internal name
is nested in a folder.

A lot of this code should be consolidated into a common ArchiveFile base
class.